requirements.txt
Django>=4.2
djangorestframework
firebase-admin
requests
python-dotenv
gunicorn
